# frontend/openapi/backend-api.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "Formato común y seguro de los errores controlados de la API.",
        "properties": {
          "code": {
            "description": "Código público y estable del error.",
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "correlationId": {
            "description": "Identificador de correlación para soporte y trazabilidad.",
            "example": "7a5f3f75-3bd2-4c47-90fc-6cfc79f1ec2d",
            "type": "string"
          },
          "message": {
            "description": "Mensaje comprensible sin detalles internos.",
            "example": "La solicitud contiene datos inválidos.",
            "type": "string"
          },
          "path": {
            "description": "Ruta de la solicitud que produjo el error.",
            "example": "/api/v1/revocation-requests",
            "type": "string"
          },
          "timestamp": {
            "description": "Fecha y hora UTC del error.",
            "example": "2026-07-20T18:30:00Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "correlationId",
          "message",
          "path",
          "timestamp"
        ],
        "type": "object"
      },
      "CurrentIdentityResponse": {
        "properties": {
          "callbackOutcome": {
            "description": "Resultado efímero del último callback, consumido una sola vez para presentación.",
            "enum": [
              "CANCELLED",
              "REJECTED",
              "IDENTITY_MISMATCH",
              "EXPIRED",
              "TIMEOUT",
              "UNAVAILABLE",
              "ERROR"
            ],
            "type": [
              "string",
              "null"
            ]
          },
          "canContinue": {
            "description": "Indica si la autorización temporal permite continuar.",
            "type": "boolean"
          },
          "nextStep": {
            "description": "Siguiente paso autorizado.",
            "enum": [
              "IDENTITY_VERIFICATION",
              "DIGITAL_CREDENTIAL_SELECTION"
            ],
            "type": "string"
          },
          "status": {
            "description": "Estado normalizado del intento de identidad.",
            "enum": [
              "STARTED",
              "VERIFIED",
              "REJECTED",
              "CANCELLED",
              "EXPIRED",
              "IDENTITY_MISMATCH",
              "ERROR"
            ],
            "type": "string"
          }
        },
        "required": [
          "canContinue",
          "nextStep",
          "status"
        ],
        "type": "object"
      },
...
      "IdentityStartResponse": {
        "properties": {
          "authorizationUrl": {
            "description": "URL de autorización construida por el backend.",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "authorizationUrl"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "FlowRefreshCookie": {
        "description": "Refresh JWT rotatorio, transportado exclusivamente mediante cookie HttpOnly.",
        "in": "cookie",
        "name": "revocacion_refresh",
        "type": "apiKey"
      },
      "FlowSessionCookie": {
        "description": "Access JWT de la sesión transaccional, transportado exclusivamente mediante cookie HttpOnly.",
        "in": "cookie",
        "name": "revocacion_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "RENIEC"
    },
    "description": "API institucional para iniciar el flujo ciudadano. La consulta pública exige Google reCAPTCHA v2 Checkbox y la autenticación usa ID Perú con continuidad temporal HttpOnly.",
    "title": "API de revocación de credenciales digitales",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/identity-verifications": {
      "post": {
        "description": "Valida la continuidad temporal, crea state y PKCE de un solo uso y devuelve la URL construida por el backend.",
        "operationId": "startIdentityVerification",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityStartResponse"
                }
              }
            },
            "description": "URL de autorización preparada"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Continuidad ausente o inválida"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La verificación no puede iniciarse en el estado actual"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Integración no disponible"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Inicia la autenticación con ID Perú",
        "tags": [
          "Verificación de identidad"
        ]
      }
    },
    "/api/v1/identity-verifications/current": {
      "get": {
        "description": "Resuelve el intento desde la cookie HttpOnly, valida la autorización y consume el resultado de presentación del callback.",
        "operationId": "getCurrentIdentityVerification",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentIdentityResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Consulta el estado de autenticación actual",
        "tags": [
          "Verificación de identidad"
        ]
      }
    },
    "/api/v1/idperu/callback": {
      "get": {
        "description": "Recibe el retorno del navegador, valida el intento y siempre redirige con HTTP 303 a una ruta frontend fija.",
        "operationId": "handleIdentityCallbackGet",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Retorno procesado y redirección controlada"
          }
        },
        "summary": "Procesa el retorno GET de ID Perú",
        "tags": [
          "Verificación de identidad"
        ]
      },
      "post": {
        "description": "Recibe el formulario del proveedor, aplica el mismo caso de uso y siempre redirige con HTTP 303 a una ruta frontend fija.",
        "operationId": "handleIdentityCallbackPost",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "error": {
                    "type": "string"
                  },
                  "session_state": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "303": {
            "description": "Retorno procesado y redirección controlada"
          }
        },
        "summary": "Procesa el retorno POST de ID Perú",
        "tags": [
          "Verificación de identidad"
        ]
      }
    },
...
  },
...
}